# Backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Fitness Tracker API
  version: 1.0.0
  description: |
    Lean Go backend for the core fitness tracking loop: users, exercise library,
    workouts, workout exercises, workout sets, meals, and weight entries.
servers:
  - url: https://localhost:8080
...
paths:
...
  /v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user
      operationId: getUser
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: User detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Users]
      summary: Update user
      operationId: updateUser
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Users]
      summary: Delete user
      operationId: deleteUser
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
...
  /v1/programs:
    get:
      tags: [Workout Programs]
      summary: List workout programs
      operationId: listPrograms
      security:
        - BearerAuth: []
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of workout programs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPrograms"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
    post:
      tags: [Workout Programs]
      summary: Create workout program
      operationId: createProgram
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgramRequest"
      responses:
        "201":
          description: Created workout program.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutProgram"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/programs/{id}:
    get:
      tags: [Workout Programs]
      summary: Get workout program
      operationId: getProgram
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Workout program detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutProgram"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Workout Programs]
      summary: Update workout program
      operationId: updateProgram
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProgramRequest"
      responses:
        "200":
          description: Updated workout program.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutProgram"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Workout Programs]
      summary: Delete workout program
      operationId: deleteProgram
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Workout program deleted.
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/programs/{id}/weeks:
    post:
      tags: [Workout Programs]
      summary: Add week to workout program
      operationId: createProgramWeek
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgramWeekRequest"
      responses:
        "201":
          description: Created program week.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramWeek"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/programs/{id}/assignments:
    get:
      tags: [Workout Programs]
      summary: List assignments for a program
      operationId: listProgramAssignmentsForProgram
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Program assignments.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgramAssignment"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
    post:
      tags: [Workout Programs]
      summary: Assign a program to a user
      operationId: createProgramAssignment
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgramAssignmentRequest"
      responses:
        "201":
          description: Created program assignment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramAssignment"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/program-weeks/{id}:
    get:
      tags: [Workout Programs]
      summary: Get program week
      operationId: getProgramWeek
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Program week detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramWeek"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Workout Programs]
      summary: Update program week
      operationId: updateProgramWeek
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProgramWeekRequest"
      responses:
        "200":
          description: Updated program week.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramWeek"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Workout Programs]
      summary: Delete program week
      operationId: deleteProgramWeek
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Program week deleted.
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/program-weeks/{id}/sessions:
    post:
      tags: [Workout Programs]
      summary: Add session to program week
      operationId: createProgramSession
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgramSessionRequest"
      responses:
        "201":
          description: Created program session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramSession"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/program-sessions/{id}:
    get:
      tags: [Workout Programs]
      summary: Get program session
      operationId: getProgramSession
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Program session detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramSession"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Workout Programs]
      summary: Update program session
      operationId: updateProgramSession
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProgramSessionRequest"
      responses:
        "200":
          description: Updated program session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramSession"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Workout Programs]
      summary: Delete program session
      operationId: deleteProgramSession
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Program session deleted.
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /v1/admin/program-assignments/{id}:
    patch:
      tags: [Workout Programs]
      summary: Admin update program assignment
      operationId: adminUpdateProgramAssignment
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProgramAssignmentRequest"
      responses:
        "200":
          description: Updated program assignment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramAssignment"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Workout Programs]
      summary: Admin delete program assignment
      operationId: deleteProgramAssignment
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Program assignment deleted.
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /v1/foods:
    get:
      tags: [Foods]
      summary: List foods
      operationId: listFoods
      parameters:
        - name: name
          in: query
          schema:
            type: string
        - name: brand
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of foods.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedFoods"
    post:
      tags: [Foods]
      summary: Create food
      operationId: createFood
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFoodRequest"
      responses:
        "201":
          description: Created food.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Food"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
...
  /v1/foods/{id}:
    get:
      tags: [Foods]
      summary: Get food detail
      operationId: getFood
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Food detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Food"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Foods]
      summary: Update food
      operationId: updateFood
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFoodRequest"
      responses:
        "200":
          description: Updated food.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Food"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Foods]
      summary: Delete food
      operationId: deleteFood
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Food deleted.
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          description: Food is in use and cannot be deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Error"
...
  /v1/admin/users:
    get:
      tags: [Admin Users]
      summary: List all users
      operationId: getAdminListUsers
      security:
        - BearerAuth: []
      parameters:
        - name: email
          in: query
          schema:
            type: string
        - name: name
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, banned, all]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: A list of users.
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/admin/users/{id}:
    get:
      tags: [Admin Users]
      summary: Get user by ID
      operationId: getAdminUser
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: A user and their stats.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  stats:
                    type: object
                    properties:
                      workouts_count:
                        type: integer
                      meals_count:
                        type: integer
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Admin Users]
      summary: Update a user
      operationId: updateAdminUser
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                role:
                  type: string
                goal:
                  type: string
                activity_level:
                  type: string
                avatar:
                  type: string
      responses:
        "200":
          description: The updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Admin Users]
      summary: Hard delete a user
      operationId: deleteAdminUser
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
        - name: confirm
          in: query
          required: true
          schema:
            type: string
            enum: ["true"]
      responses:
        "204":
          description: User successfully deleted.
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/admin/users/{id}/ban:
    post:
      tags: [Admin Users]
      summary: Ban a user
      operationId: banAdminUser
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: The banned user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/admin/users/{id}/unban:
    post:
      tags: [Admin Users]
      summary: Unban a user
      operationId: unbanAdminUser
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: The unbanned user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/admin/dashboard/summary:
    get:
      tags: [Admin Dashboard]
      summary: Get executive dashboard summary
      operationId: getAdminDashboardSummary
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Admin dashboard summary.
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/admin/dashboard/trends:
    get:
      tags: [Admin Dashboard]
      summary: Get activity trends
      operationId: getAdminDashboardTrends
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of daily activity stats.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
  /v1/admin/users/growth:
    get:
      tags: [Admin Dashboard]
      summary: Get user growth stats
      operationId: getAdminUserGrowth
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Daily new user counts.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/admin/workouts/stats:
    get:
      tags: [Admin Dashboard]
      summary: Get workout analytics
      operationId: getAdminWorkoutStats
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Workout analytics.
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
  /v1/admin/nutrition/stats:
    get:
      tags: [Admin Dashboard]
      summary: Get nutrition analytics
      operationId: getAdminNutritionStats
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Nutrition analytics.
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/admin/foods:
    get:
      tags: [Admin Dashboard]
      summary: List foods with admin metadata
      operationId: getAdminFoods
      security:
        - BearerAuth: []
      parameters:
        - name: name
          in: query
          schema:
            type: string
        - name: brand
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of foods with admin-only usage metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAdminFoods"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/admin/moderation/stats:
    get:
      tags: [Admin Moderation]
      summary: Get moderation analytics
      operationId: getAdminModerationStats
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Moderation analytics.
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    ID:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NoContent:
      description: Resource deleted successfully.
  schemas:
...
    User:
      type: object
      required:
        - id
        - email
        - name
        - role
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        avatar:
          type: string
        role:
          type: string
          enum: [user, admin]
        age:
          type: integer
          deprecated: true
          description: Deprecated. Use date_of_birth instead.
        date_of_birth:
          type: string
          format: date
          nullable: true
        weight:
          type: number
          format: double
        height:
          type: number
          format: double
        goal:
          type: string
        activity_level:
          type: string
        tdee:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT